Move most functionality from AuditLogMessage to the base class
